Update CoordinatorTests to ES patch level/ introduce AbstractCoordinatorTestCase
The CoordinatorTests were flaky because they missed some adjustements
that were made upstream in AbstractCoordinatorTestCase.

This brings the files closer to the ES patch level we are at.
